Add mongodb and word counter
Now you can upload text, which will be analysed and afterwards
stored in a mongodb, which is running in a seperate docker container.
The Analysis of the input text is a simple word counter.
The information is safed in the database, too.
Afterwards all the stored information can be seen in a table from vuejs.